Moved test modules to their own files
