+++ boa_cli/Cargo.toml
@@ -11,5 +11,5 @@ exclude = ["../.vscode/*", "../Dockerfile", "../Makefile", "../.editorConfig"]
 edition = "2018"
 
 [dependencies]
-Boa = { path = "../boa" }
+Boa = { path = "../boa", default-features = false }
 structopt = "0.3.9"
